pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.0.0"
description = "Публичная версия бота для Discord"
authors = ["GREEN Corporation <vtcgreenlogistic@gmail.com>"]
license = "GPL-2.0"
readme = "README.md"
packages = [{include = "discord_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.1.0"
asyncio = "^3.4.3"
logging = "^0.4.9.6"
pytest = "^7.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
